Address remaining feedback
Thanks @mats-sx
