+++ spark-cypher/src/test/scala/org/opencypher/spark/impl/acceptance/CAPSScanGraphAcceptanceTest.scala
@@ -15,261 +15,8 @@
  */
 package org.opencypher.spark.impl.acceptance
 
-import org.opencypher.okapi.api.schema.{PropertyKeys, Schema}
-import org.opencypher.okapi.api.types.{CTInteger, CTString}
-import org.opencypher.okapi.api.value.CypherValue
-import org.opencypher.okapi.api.value.CypherValue.CypherMap
-import org.opencypher.okapi.logical.api.configuration.LogicalConfiguration.PrintLogicalPlan
-import org.opencypher.okapi.relational.api.configuration.CoraConfiguration.PrintPhysicalPlan
-import org.opencypher.spark.impl.CAPSConverters._
-import org.opencypher.spark.impl.CAPSPatternGraph
 import org.opencypher.spark.test.support.creation.caps.{CAPSScanGraphFactory, CAPSTestGraphFactory}
-import org.opencypher.spark.schema.CAPSSchema._
-
-import scala.collection.Bag
 
 class CAPSScanGraphAcceptanceTest extends AcceptanceTest {
   override def capsGraphFactory: CAPSTestGraphFactory = CAPSScanGraphFactory
-
-  def testGraph1 = initGraph("CREATE (:Person {name: 'Mats'})")
-
-  def testGraph2 = initGraph("CREATE (:Person {name: 'Phil'})")
-
-  def testGraph3 = initGraph("CREATE (:Car {type: 'Toyota'})")
-
-  it("should construct a graph") {
-    val query =
-      """|CONSTRUCT {
-         |  CREATE (:A)-[:KNOWS]->(:B)
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.labels should equal(Set("A", "B"))
-    result.getGraph.schema.relationshipTypes should equal(Set("KNOWS"))
-    result.getGraph.nodes("n").iterator.length should equal(2)
-  }
-
-  it("should CONSTRUCT a graph with multiple connected CREATE clauses") {
-    val query =
-      """|CONSTRUCT {
-         |  CREATE (a:A)-[:KNOWS]->(b:B)
-         |  CREATE (b:B)-[:KNOWS]->(c:C)
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.labels should equal(Set("A", "B", "C"))
-    result.getGraph.schema.relationshipTypes should equal(Set("KNOWS"))
-    result.getGraph.nodes("n").iterator.length should equal(3)
-  }
-
-  it("should CONSTRUCT a graph with multiple unconnected CREATE clauses") {
-    val query =
-      """|CONSTRUCT {
-         |  CREATE (a:A)-[:KNOWS]->(b:B)
-         |  CREATE (c:C)-[:KNOWS]->(d:D)
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.labels should equal(Set("A", "B", "C", "D"))
-    result.getGraph.schema.relationshipTypes should equal(Set("KNOWS"))
-    result.getGraph.nodes("n").iterator.length should equal(4)
-  }
-
-  it("should CONSTRUCT a graph with multiple unconnected anonymous CREATE clauses") {
-    val query =
-      """|CONSTRUCT {
-         |  CREATE (:A)
-         |  CREATE (:B)
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.labels should equal(Set("A", "B"))
-    result.getGraph.schema.relationshipTypes should equal(Set.empty)
-
-    result.getGraph.nodes("n").size should equal(2)
-    result.getGraph.relationships("r").size should equal(0)
-  }
-
-  it("should construct a node property from a matched node") {
-    val query =
-      """|MATCH (m)
-         |CONSTRUCT {
-         |  CREATE (a :A)
-         |  SET a.name = m.name
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.labels should equal(Set("A"))
-
-    result.getGraph.schema should equal(Schema.empty.withNodePropertyKeys("A")("name" -> CTString).asCaps)
-
-    result.getGraph.cypher("MATCH (a:A) RETURN a.name").getRecords.iterator.toBag should equal(Bag(
-      CypherMap("a.name" -> "Mats")
-    ))
-  }
-
-  it("should construct a node property from a literal") {
-    val query =
-      """|CONSTRUCT {
-         |  CREATE (a :A)
-         |  SET a.name = 'Donald'
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.labels should equal(Set("A"))
-
-    result.getGraph.schema should equal(Schema.empty.withNodePropertyKeys("A")("name" -> CTString).asCaps)
-
-    result.getGraph.cypher("MATCH (a:A) RETURN a.name").getRecords.iterator.toBag should equal(Bag(
-      CypherMap("a.name" -> "Donald")
-    ))
-  }
-
-  it("should construct multiple properties") {
-    val query =
-      """|CONSTRUCT {
-         |  CREATE (a :A:B)
-         |  SET a.name = 'Donald'
-         |  SET a.age = 100
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.labels should equal(Set("A", "B"))
-
-    result.getGraph.schema should equal(Schema.empty.withNodePropertyKeys(Set("A", "B"), PropertyKeys("name" -> CTString, "age" -> CTInteger)).asCaps)
-
-    result.getGraph.cypher("MATCH (a:A) RETURN a.name").getRecords.iterator.toBag should equal(Bag(
-      CypherMap("a.name" -> "Donald")
-    ))
-  }
-
-  it("should pick up labels of the outer match") {
-    val query =
-      """|MATCH (m:Person)
-         |CONSTRUCT {
-         |  CREATE (a~m)
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.labels should equal(Set("Person"))
-
-    result.getGraph.schema should equal(Schema.empty.withNodePropertyKeys(Set("Person"), PropertyKeys("name" -> CTString)).asCaps)
-
-    result.getGraph.cypher("MATCH (a:Person) RETURN a.name").getRecords.iterator.toBag should equal(Bag(
-      CypherMap("a.name" -> "Mats")
-    ))
-  }
-
-  it("should construct a relationship") {
-    val query =
-      """|CONSTRUCT {
-         |  CREATE ()-[r:FOO]->()
-         |  SET r.val = 42
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.relationshipTypes should equal(Set("FOO"))
-
-    result.getGraph.schema should equal(Schema.empty
-      .withNodePropertyKeys()()
-      .withRelationshipPropertyKeys("FOO", PropertyKeys("val" -> CTInteger)).asCaps)
-
-    result.getGraph.cypher("MATCH ()-[r]->() RETURN r.val").getRecords.iterator.toBag should equal(Bag(
-      CypherMap("r.val" -> 42)
-    ))
-  }
-
-  it("should tilde copy a relationship") {
-    val query =
-      """|CONSTRUCT {
-         |  CREATE ()-[r:FOO]->()
-         |  SET r.val = 42
-         |}
-         |MATCH ()-[s]->()
-         |CONSTRUCT {
-         |  CREATE ()-[t~s]->()
-         |  SET t.name = 'Donald'
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.relationshipTypes should equal(Set("FOO"))
-
-    result.getGraph.schema should equal(Schema.empty
-      .withNodePropertyKeys()()
-      .withRelationshipPropertyKeys("FOO", PropertyKeys("val" -> CTInteger, "name" -> CTString)).asCaps)
-
-    result.getGraph.cypher("MATCH ()-[r]->() RETURN r.val, r.name").getRecords.iterator.toBag should equal(Bag(
-      CypherMap("r.val" -> 42, "r.name" -> "Donald")
-    ))
-  }
-
-  // TODO: Support MGC syntax
-  ignore("should allow simple MGC syntax") {
-    val query =
-      """|CONSTRUCT {
-         |  CREATE (a:A)-[r:FOO]->(b:B)
-         |}
-         |MATCH (a)-->(b)
-         |CONSTRUCT {
-         |  CREATE (a)-[:KNOWS]->(b)
-         |}
-         |RETURN GRAPH""".stripMargin
-
-    val result = testGraph1.cypher(query)
-
-    result.getRecords.toMaps shouldBe empty
-
-    result.getGraph.schema.relationshipTypes should equal(Set("KNOWS"))
-    result.getGraph.schema.labels should equal(Set("A", "B"))
-
-    result.getGraph.schema should equal(Schema.empty
-      .withNodePropertyKeys("A")()
-      .withNodePropertyKeys("B")()
-      .withRelationshipPropertyKeys("KNOWS")().asCaps)
-
-    result.getGraph.cypher("MATCH ()-[r]->() RETURN type(r)").getRecords.iterator.toBag should equal(Bag(
-      CypherMap("type(r)" -> "KNOWS")
-    ))
-  }
-
 }
